feat: refactor data structure
