+++ packages/core/index.ts
@@ -1,4 +1,5 @@
 export * from './useGoogleMap'
 export * from './useInfoWindow'
+export * from './useMap'
 export * from './useMarker'
 export * from './useSearch'
